+++ ruby-sinatra/app.rb
@@ -5,14 +5,12 @@ require 'sinatra/activerecord/rake'
 require 'mysql2'
 require 'json'
 
-require './routes'
-
 class App < Sinatra::Application
 end
 
-require_relative 'helpers/init'
-
 # Require models
-current_dir = Dir.pwd
-Dir["#{current_dir}/models/*.rb"].each { |file| require file }
+Dir["./app/models/*.rb"].each { |file| require file }
+Dir["./app/helpers/*.rb"].each { |file| require file }
+Dir["./app/controllers/*.rb"].each { |file| require file }
 
+# App.helpers ApplicationHelper
