+++ ruby-sinatra/config.ru
@@ -1,0 +1,2 @@
+require './app'
+run Sinatra::Application
