--- ruby-sinatra/helpers/init.rb
@@ -1,2 +1,0 @@
-require_relative 'application_helper'
-App.helpers ApplicationHelper
